Moved windows definitions to folder
