+++ .gitignore
@@ -1,1 +1,5 @@
 .vscode/settings.json
+.vscode
+__pycache__
+szch.mat
+.gitignore
