+++ main.py
@@ -1,7 +1,7 @@
 import sys
 import numpy as np
 from PyQt5.QtWidgets import QApplication, QMainWindow, QSizePolicy
-from GraphDisplay import Ui_MainWindow
+from windows.GraphDisplay import Ui_MainWindow
 
 from SignalProcesing import Math, Function
 from NewSignal import NewFunctWindow
