+++ windows/NewSignalBox.ui
